Moved random network enums into own file
