+++ Key Conservation/Key Conservation/Models/UserController.swift
@@ -8,18 +8,9 @@
 
 import Foundation
 
-struct Bearer: Codable {
-    let token: String
-}
-
-enum LoginType: String {
-    case signUp = "signup"
-    case signIn = "login"
-}
-
 class UserController {
     var bearer: Bearer?
-    private let baseURL = URL(string: "https://lambdagigs.vapor.cloud/api")!
+    private let baseURL = URL(string: "https://")! //placeholder
     
     func loginWith(user: User, loginType: LoginType, completion: @escaping (Error?) -> ()) {
         let requestURL = baseURL.appendingPathComponent("users/\(loginType.rawValue)")
@@ -66,6 +57,10 @@ class UserController {
             
             completion(nil)
             }.resume()
+    }
+    
+    func updateInfo(for user: User, completion: @escaping (NetworkError?) -> ()) {
+        // Stub function
     }
